Add cherry-pick bot for the pulsar (#8500)
**Motivation**

We want to make the pulsar can cherry-pick automatically. This action will make a closed PR can cherry-pick by the command /pulsarbot cherry-pick to branch-X.Y.

When a PR is merged, the bot will say,

Hey. If you want to cherry-pick this pr to a target branch, please comments '/pulsarbot cherry-pick to branch-X.Y'.
Then you can comment /pulsarbot cherry-pick to branch-X.Y, the bot will open a new PR for cherry-picking the closed PR.

If we open a new PR to do the cherry-pick, we can also use the CI to check the branch.