Support namespace-level duplication snapshot (#8506)
Master Issue: #8237

### Motivation
Currently we take de-duplication snapshots based on size. If the topic has relatively low traffic, the de-duplication cursor will not move. This can cause messages that are not able to be deleted based on the retention policy. We should add a policy to take de-duplication snapshots based on time.

### Modifications
add api to get/set/remove namespace-level policies

### Verifying this change
TopicDuplicationTest.java