[Transaction] Fix the Github CI Test (#8541)
Motivation
Try to fix the Github CI.

## implement
because we consume all transaction messages than have been produced. but commit marker may not delete, because when a commit marker operation is processing, when we ack the position will not allow to do delete operation of commit marker, so the commit marker may not delete. we should handle this case.