Fix the pulsar sql can not query the last message issue (#8635)
Fixes #7811

### Motivation

Currently, the bookkeeper client gets the LAC is the piggyback lac which is carried by the next message. This causes the pulsar SQL connector will always get the `message number -1` messages. If we want to query all messages, we need to use the explicit LAC to get the total number of entries. 
BookKeeper 4.12.0 makes us can use the `useV2Protocol` to enable using v3 protocol which will get the explicit LAC.

### Modifications

- Add the properties for the pulsar SQL.
- Update the integration test to query the last message.