8704-improve env config handling
There are three issues addressed in this PR.
1) script files that contain lines other than comments and value
assignments cause the script to abort not re-writing the file at all.
These lines should be tolerated if they cannot be parsed as value
assignments and flow through to the edited results.

2) Redaction of passwords in environment variables. Environment
variables with a name containing password will automatically be
redacted and not printed in the log messages.

3) values that contain spaces as in many java args that contain
stringed together options like "-Xmx ... -Xmn ..." are read from the
environment without the quotes. These quotes need to be added back
when writing back to the shell script to ensure that these values are
processed as a group.