fix flaky-test: TopicsConsumerImplTest.testConsumerUnackedRedelivery (#8738)
Fixes #8710

### Modifications
Re-delivery of the message takes time. Because the re-delivery uses a timer, it is not accurate. The timeout period for receiving messages in the unit test is too short, and it is easy to time out. When we shorten the timeout period from 100ms to 10ms, this problem can be easily reproduced.

Therefore, we extend the timeout period to 2 seconds