Fix the problem that batchMessageId is converted to messageIdImpl (#8779)
Fixes #8712

### Motivation
TopicReaderTest.testHasMessageAvailable is flaky
Cause Analysis:
When there is only one message in the batch, calling Broker's `GetLastMessageId` will determine that this is not a batch message, and then return a `MessageIdImpl`. (See ServerCnx line 1553)
In the same scenario, the client will return a `BatchMessageId` after sending a message. (See ProducerImpl line 1151)

This will happen:
MessageIdImpl: `3:31:-1`
BatchMessageId: `3:31:-1:0`

When calling `reader.hasMessageAvailable()`, the two ids will be compared, like this: `lastMessageIdInBroker.compareTo(messageId)`
Although it is the same messageId, the results will not be equal.

### Modifications
When we call `admin.topics().getLastMessageId`, even if there is only one message in the batch, it is considered to be BatchMessageId. Client also acts like this.
Therefore, we keep consistent everywhere.
Even if there is only one message in the batch, we think it is `BatchMessageId`


(cherry picked from commit 85f3ff4edbaa10c7894af8ad823cbce37b13829c)