Fix single-quotes added to user-conf (#8780)
Fixes #8757

### Motivation

For the config content of Go Function, different runtime types require different parameters. Under Process Runtime, we only need the json string of config content itself; for k8s runtime, since we need to use the config content object to splice the command line in the pod, we need to add single quotes to the config content for processing.

### Modifications

In this pull request, we introduced a `k8sRuntime` flag label, which is used to identify different runtime forms.