[Doc]--Update docs for how to delete data from namespace for releases 2.7.0--2.6.0 (#8934)
### Motivation


This PR is a follow-up PR for #8760, in which only master doc is updated. This PR is used to sync doc updates for releases 2.6.0--2.7.0, including minor releases.

Updates in master release have been reviewed and approved. You can just approve it directly. 

### Modifications

Add a section in cook-retention policy for deleting data from namespace.