Fix the metadata setup compatibility issue (#8959)
---

*Motivation*

metadata command is not compatible with old versions.
For keeping the compatibility of the setup metadata command,
we should avoid deleting the existing arguments. If we need to
change it, it's better to keep the old arguments and mark it
deprecates and hide them.

(cherry picked from commit d7da7af13cb86cd5684434c2fe4242d132cbd107)