Make function supports to use package command (#8973)
---

Master Issue: #8676

*Motivation*

We have a new package management service which can manage all
the packages. We can use that in the Pulsar Function will make the 
function packages are easier to manage.
I reuse the 'packageUrl' for downloading the package from the packages
management service. So user can use the package name as the 'packageUrl'
to download it.

*Modifications*

- Support new package type URL for creating and updating functions
- Support to use package download command for Kubernetes runtime
- Add tests for the changes