Support create and update source with package name (#8988)
---

Master Issue: #8676

*Motivation*

We have a new package management service that can manage all
the source packages. We can use that in the Pulsar Source to manage
source packages.
I reuse the 'packageUrl' for downloading the package from the packages 
management service. So user can use the package name as the 'packageUrl'
to download it.

*Modifications*

- Support create and update the source with package name
- Add test for this