Support create and update sink with package name (#8987)
---

Master Issue: #8676

*Motivation*

We have a new package management service that can manage all
the sink packages. We can use that in the Pulsar Sink to manage
sink packages.
I reuse the 'packageUrl' for downloading the package from the packages
management service. So user can use the package name as the 'packageUrl'
to download it.

*Modifications*

- Support create and update sink with the package name
- Add test for this