Additional error checks in TwoPhasesCompactor (#9133)
If there is an exception in the `whenComplete()` block, the exception is "eaten" up and the `loopPromise` is never set to failed.

Instead, it's better to use the `thenAccept().exceptionally()` to ensure that any exception in the first handler will be propagated back.

(cherry picked from commit 984bf837489b7c6e5024d42a74fd729461baaf1e)