Issue 9082: Broker expires messages one at a time after topic unload (#9083)
Fixes #9082

In case of topic unload the ManagedLedger rolls a new ledger and this confuses the search for messages to be expired.

This is how OpFindNewest (the operation involved in message expiry, PersistentMessageExpiryMonitor) works:
- first check if first message is expired
- test in the last message
- perform a search...

At step two we are jumping to a position that is not valid, and the search ends immediately, returning only the first message

Ensure that we are jumping only to a valid position, this change allows the PersistentMessageExpiryMonitor to find the best range of messages to expire.

This change added tests:
- unit test that covers the change
- one reproducer for the problem, that now is fixes
- another test case for the expected end-to-end behaviour for a consumer that waits too much and messages expire

It can be also verified trying to reproduce manually the reproducer at #9082, after applying this patch the issue is no more reproducible

(cherry picked from commit 7313b3403867d5ca16ea68edeaedd42f8c389ad3)