[Docs]Add documentation for consumerless subscriptions and their modes (#9125)
### Motivation

This PR seeks to clarify subscriptions without consumers by adding new documentation. In https://github.com/apache/pulsar/issues/8906, @merlimat helped me understand these subscriptions, and this PR adds that knowledge to the documentation. [Here](https://github.com/apache/pulsar/issues/8906#issuecomment-742836792) is the specific comment that provides the basis for this PR.

### Modifications

Adds documentation in the subscription mode section to describe how subscription modes are defined and that they are undefined when there aren't consumers.

This is only a documentation change.