[issue 8337][Worker] Move initialize dlog namespace metadata to bin/pulsar (#8781)
Fixes #8337 

### Motivation

Currently, the starting function worker service will initialize the distributed log namespace. when initializing the distributed log namespace, the zookeeper will be connected. A better way is to implement a tool to initialize the distributed log namespace, which is similar to `./bin/pulsar initialize-cluster-metadata`.

### Modifications

- Add init distributed log namespace metadata to `./bin/pulsar initialize-cluster-metadata`
- Add flag to control whether the distributed log namespace is initialized in Function worker
- Add flag to `conf/functions_worker.yml`

### Doc

If you want to initialize distributed log metadata by `bin/pulsar`, you need to perform the following steps:
1. Using `./bin/pulsar initialize-cluster-metadata` to initialize cluster
2. Set initializedDlogMetadata to `true` in `functions_worker.yml`

**Note**: All the changes are compatible.