Separate the Functions validation and transforamtion ---
Fixes #9180

*Motivation*

When creating a function, we will do the validation for the
function package and transform the FunctionConfig to the
FunctionDetail in the same method.
We introduced package management for saving the function
packages, and want to reduce the duplicated validation when
creating a function with a package name.
So I suggest we need to separate the validation steps and
transformation step to make the method more clear.