Fix potential http get hangs in the Pulsar Admin (#9203)
### Motivation

Currently, for the HTTP async get method, we pass a callback and return a future. This is confusing which one is the right result. And if exception throws, will return a future that complete with exception but the callback have not been called. All the caller of the `asyncGetRequest` handles the callback, not the returned future, this will lead a potential hangs in the Pulsar Admin since the callback might never be called.

### Modifications

Change the `asyncGetRequest` method to only handle the callback and use `void` as the return type.