Handle web application exception to redirect request (#9228)
*Motivation*

`validateNamespaceBundleOwnership` throws a web application exception with `Response.temporaryRedirect(redirect)`
to redirect the request to the owner broker. But the web application exception is a runtime exception. If you don't
handle and propagate it correctly, it will cause web request to be hang.

PR #8746 changed some web calls to async but it doesn't handle the web application exception. It causes `topics list`
with non-persistent topics to be hang.

This pull request make sure the callers of `validateNamespaceBundleOwnership` catch and propagate the exceptions to
web response.

(cherry picked from commit aaa27c3fd2dfe75669564ccc00c50a9aed279915)