Fixing the mac build if you use a different openssl (#9165)
Fixes #9155

### Motivation
When I attempted to build the c library on macosx with openssl 1.1.1i it failed with linkage errors. I tracked those errors down to the fact that the OPENSLL_ROOT_DIR was incorrect outside of mac vms used by the upstream.

### Modifications

I set the default value first and then reset it only if running on mac.

### Verifying this change

This change is a trivial rework / code cleanup without any test coverage.