[Functions] Can't create functions with m-TLS
*Motivation*

apache/pulsar#7255 re-worked Function MetaDataManager to make all metadata writes only by the leader.
This unintentionally broke Pulsar Functions when m-TLS is used for authentication. Because it doesn't
taken TLS port into consideration and always uses a non-TLS port to communicate with the leader broker.

The PR fixes the broken implementation and ensure Pulsar Functions use the right service url and
authentication plugin to communicate with leader.

*Tests*

Add an integration test to reproduce the issue and ensure functions worker with m-TLS