PLSR-1470: Only auth _errors_ should log at error level
An authentication failure should not log at error level, as it depends
on the credentials that the client provides, which may well be wrong.

An authentication/authorization error, where it cannot be decided if
credentials are valid or not, is an error that needs operator
attention and therefore should be logged at error level.