Support fetching metadata from entry data in publish callback (#9257)
### Motivation

[PIP-70](https://github.com/apache/pulsar/wiki/PIP-70%3A-Introduce-lightweight-broker-entry-metadata) introduced lightweight broker entry metadata which is added by related interceptor like `AppendBrokerTimestampMetadataInterceptor` before being written to bookie asynchronously. However, in the callback when entries are written to bookie successfully, the caller side can only get the ledger id and entry id, see `PublishContext#completed`:

```java
void completed(Exception e, long ledgerId, long entryId);
```

Currently there's no way to get the broker entry data in publish callback. We can access the interceptor's internal field but it may be overwritten by the next entry's broker entry metadata because the interceptor changes the internal fields when the entry was sent while the callback is invoked when the send is done.

This PR intends to expose the entry's data to callbacks, including managed ledger's add callback and the publish context's complete callback. If we want to make use of broker entry metadata or the old message metadata in future, we can use the new callbacks.

### Modifications

- Add an extra `ByteBuf` argument to `AddEntryCallback#addComplete` to expose the entry data.
- Add a default method `setMetadataFromEntryData` to `PublishContext` interface to allow implementer side retrieve some metadata from entry data.
- Add tests for `AddEntryCallback#addComplete`.
- Add tests for `PublishContext#setMetadataFromEntryData`.