Using the same job to run cancel workflow ---
**Motivation**

After introducing the new cancel CI workflow, it takes
a lot of job place which cause we have more and more CI
in the queue. I saw how the apache/airflow uses, the
apache/airflow run it as a step, so let's remove it into
the separate workflows. It will reduce our CI job usage.
Also, I test it in my own repo, it looks like working well.
https://github.com/zymap/pulsar/actions?query=branch%3Atest-duplicate-ci'