Allow to configure BookKeeper Opportunistic Striping Feature (and all BK client features using bookkeeper_ prefix) (#9232)
BookKeeper 4.12 introduces the 'Opportunistic Striping' feature.
In BK terms 'striping' happens when EnsembleSize is greater than WriteQuorumSize, in this mode the entries are distributed round robin over a set of bookies, in order to  achieve better performances as you can use the resources of more bookies.

For instance in a small HA cluster, with only 3 bookies, you must run Pulsar with 2-2-2 replication parameters (EnsembleSize=2,WriteQuorumSize=2,AckQuorumSize=2).
You cannot set EnsembleSize=3 (and thus use 'striping') because in case of temporary outage of a single bookie the BK client is not able to create an ensemble with 3 bookies.

With Opportunistic Striping you can use 3-2-2 and when the system is fully up-and-running with 3 bookies then you go with striping, but during single bookie outages (like during reconfigurations/updates) you fall back to 2-2-2.

This is not about consistency or durability, it is only about having the ability to get the most out of your bookkeeper cluster.

- Add a generic way to configure internal BookKeeper client options, any entry start starts with `bookkeeper_` is passed to the BK client configuration after stripping the prefix
- Add unit tests for `bookkeeper_opportunisticStriping` configuration option
- Add configuration example in broker.conf

- the change add new tests

(cherry picked from commit bc69ad29e01a005ef2dc3eae36aed429b42495aa)