[client] Cleanup consumer on multitopic subscribe failure
Currently, when a multi-topic subscribe fails (via a set of topics or a
regex) we can leave consumers connected, as the multitopic consumer
doesn't close any of the topics.

This means we rely on the client to call closeAsync, otherwise, the
consumer is left in partially open state.

This fix changes that, and ensures we call close in the case of an
exception