Do not use a static map of listeners in TopicPoliciesService (#9486)
Maybe CI jobs are failing with OOM in the brokers unit tests. The Surefire worker is configured with 4 processes, each with xmx of 1G.

The problem was introduced in #7863 where a static map of listeners was added to an interface. That makes that map to contain all the `PulsarService` instances created during the tests execution and keeping references to everything else.

The map should instead be scoped to the specific instance.

(cherry picked from commit 31ee4541a16bd7194d50b576434eb7742ef87e99)