Fix flaky test ConsumedLedgersTrimTest.testConsumedLedgersTrimNoSubscriptions (#9420)
Fixes #9409
### Motivation
After restarting, it takes time to load the topic, and this admin api will throw an exception when the topic cannot be queried.
This problem can easily be triggered when this unit test is called in a loop. Therefore, I added a judgment to determine whether the topic is loaded. After that, I called it 100 times in a loop and no exception was seen.