Fixes canceling of duplicate workflow runs
The previous way of canceling workflow runs was flawed because
it was run as part of `pull_request` workflows - which did not
work for fork PRs. The reason is, that `pull_request` workflows
are executed with "read-only" `GITHUB_TOKEN`. Therefore running
the action within the PR ends with 'resource not accessible by
integration` error.

This PR implements cancel action in the way that it was intended:
the 'cancel-workflow-run' action runs within a `workflow_run`
triggered workflow which has two characteristics:

* it runs only using master version of workflow (so this workflow
  will become active only after we merge it to master)

* it runs using 'write' GITHUB_TOKEN - this allows actually to
  cancel the workflows withot the 'resource...' error.

In order to optimise a number of machines/jobs started, there is
only one 'CI - Cancel duplicate workflows' workflow, which
has one job with many steps - each steps cancels duplicates in a
different workflow. This `workflow_run` is triggered by the
`CI - Unit` workflow, which has been somewhat arbitrary chosen - but
this could be any of the workflows that runs with every PR.

Runing this as a single workflow/job has several benefits:

* only one job is started and takes less of the queue
* one machine is initialized once for all steps thus the overhead
  of creating machine is much smaller
* splitting workflows to separate steps has the benefit of seeing
  each cancelling action separately

The reason why it is done as copy&paste rather than matrix or
yaml anchor are simple:

* matrix creates sepearate job for each matrix entry - increasing
  the job queue as well as adding overhead of machine creation
* Github Action does not support yaml anchors

All the `cancel` actions run with `allDuplicates` cancel mode
which causes them to agressively kill all duplicates even in
high-strain queu case - the first of the cancel jobs that starts
running will activelly kill all the duplicates of all the
workflows from all PRs running in the repository.