Fix create partition of an exist topic doesn't throw RestException (#9342)
### Motivation

Currently creating a partition of an existed partitioned topic was created doesn't throw any exception. However it should be an invalid behavior. The reason is that when a non partitioned topic was created, it only checks whether the number of partitions is positive. However, no matter the topic doesn't exist or the topic is an existed partition, the number of partitions is 0. This PR is to distinguish these two cases and throw a `RestException` when the non-partitioned topic is an existed partition.

### Modifications

- When creating a non-partition topic that is an existed partition, throw a `RestException` and add a test to verify it.
- Add a test provider to ensure backward compatibility that users can still create missed partitions by creating non-partitioned topics of the missed partition.
- Fix the functions worker startup. Before this PR it always tries to create the metadata topics like `public/functions/assignment` no matter if it exists. Here we ignore the `ConflictException`.

### Verifying this change

This change added tests and can be verified as follows:

  - *Added PersistentTopics#testCreateExistedPartition*
  - *Added another test provider to PartitionCreationTest#testCreateMissedPartitions*
  - *Add tests for creating existed topics to `AdminApiTeststestPersistentTopicCreation`.*