[Issue9507][testclient] add --batch-index-ack for the pulsar-perf (#9521)
Fixes #9507 

### Motivation

PIP-45 https://github.com/apache/pulsar/wiki/PIP-54:-Support-acknowledgment-at-batch-index-level introduced batch index level acknowledge but the pulsar-perf client does not support to enable it for the consumer. It's better to allow users to do the performance test for the batch index acknowledgment.

### Modifications

add --batch-index-ack for the pulsar-perf consume command and keep it as false by default.