[Issue 9480][pulsar-io] add option for auto.offset.reset to kafka source (#9482)
Fixes #9480 

### Motivation

The kafka source sets auto.offset.reset to "earliest". This means all old messages from kafka are produced to pulsar. Often is it desirable to start form the present location "latest".
The option is set after the user config has been loaded so it cannot be changed:
[source code link](https://github.com/apache/pulsar/blob/master/pulsar-io/kafka/src/main/java/org/apache/pulsar/io/kafka/KafkaAbstractSource.java#L87)

### Modifications

Added an autoOffsetReset option to KafkaSourceConfig