PLSR-1470: Only auth _errors_ should log at error level (#9325)
An authentication failure should not log at error level, as it depends
on the credentials that the client provides, which may well be wrong.

An authentication/authorization error, where it cannot be decided if
credentials are valid or not, is an error that needs operator
attention and therefore should be logged at error level.

Co-authored-by: Ivan Kelly <ikelly@splunk.com>
Co-authored-by: Matteo Merli <mmerli@splunk.com>
Co-authored-by: Sijie Guo <sijie@apache.org>
(cherry picked from commit fd9875c41bdb86afad07413fdeedcc62781e92d2)