[Compression] Fix ByteBuffer allocate error in the AirliftUtils (#9667)
Fixes #9666

### Motivation

The compressed data length may be bigger than the original data length (e.g. the source text is not repeated, such as "abcde"), so we can't use the uncompressed length as the allocated length to initial the ByteBuffer.

### Modifications

Use the capacity of the ByteBuffer instead of the uncompressed length as the allocated length.

### Verifying this change

Update the existing tests.