[FlakyTest]Try to fix flaky test ReplicatorTest.testReplicatorOnPartitionedTopic. (#9659)
Fixes #9457

### Motivation
Not sure why state is not cleaned up but exception is indicating metadata of namespace the test trying to create already exist so use a different namespace name for different test run should able to fix it.

### Modifications
Append System.nanoTime to the namespace used to avoid being affect by previous state. This seems working fine for other test cases in the same test.