Fix deleteTransactionMarker memory leak (#9751)
Fixes (part of) #9725 (not sure if there are other leaks, though)

### Motivation

Fixes a memory leak for transactions.

### Modifications

The `ReadEntryCallback` should release the `entry` passed to it. I validated that other implementations of the `ReadEntryCallback` class release the `entry`. Further, the final stack trace referenced in #9725 has clear references to this stack. The only change is to call `entry.release()`. Because parsing could lead to an exception, I used a `finally` block.

### Verifying this change

This change is a trivial rework / code cleanup without any test coverage. (We must not have test coverage for this type of leak, though.)