Fix marking individual deletes as dirty (#9732)
* Fix marking individual deletes as dirty

When we mark cursors as dirty, we aren't marking when individual acks
cause a dirty cursor.

This results in cursors not being flushed and causing redelivery.

This one line fix will ensure we mark the cursor as dirty in this
situation as well

* add a test

* improve tests to not use sleep

* make the polling rate be slower

(cherry picked from commit 34ca8938ca13ea60b05d164c27d9755855caf87c)